Tidy up some STAmount and base_uint operations: * Constrain use of arithmetic operators in STAmount * Prevent constructor conversion of uint256 to uint128 - make intent clear * Prevent implicit conversion of uint64_t to uint256 * Prevent implicit conversion of uint64_t to uint160

+++ src/ripple_app/contracts/ScriptData.h
@@ -70,7 +70,7 @@ public:
     }
     virtual uint160 getUint160 ()
     {
-        return (0);
+        return uint160(0);
     }
 
     //virtual bool isCurrency(){ return(false); }
